fix tut realz. bugs

+++ 11-Tutorium/EquilateralTriangle.java
@@ -1,5 +1,5 @@
 public class EquilateralTriangle extends RegularPolygon {
-   public EquilateralTriangle(int sideLength) {
+   public EquilateralTriangle(double sideLength) {
       super(3, sideLength);
    }
